Refactoring del codice
