+++ pokemon-vgc-engine-master/example/Example_Competitor.py
@@ -1,14 +1,15 @@
 from vgc.behaviour import BattlePolicy, TeamSelectionPolicy, TeamBuildPolicy
-from vgc.behaviour.BattlePolicies import RandomPlayer, MyPolicy
+from vgc.behaviour.BattlePolicies import RandomPlayer
 from vgc.behaviour.TeamBuildPolicies import TerminalTeamBuilder, RandomTeamBuilder
 from vgc.behaviour.TeamSelectionPolicies import FirstEditionTeamSelectionPolicy
 from vgc.competition.Competitor import Competitor
-
 from vgc.behaviour import BattlePolicy, TeamSelectionPolicy, TeamBuildPolicy
-from vgc.behaviour.BattlePolicies import RandomPlayer, MyPolicy, MyMinimax, MyMinimaxWithAlphaBetaKiller, MyMinimaxWithAlphaBetaSortedKiller,MyMinimaxWithAlphaBetaKillertransposition,MCTS_MS,AggressiveMinimaxWithMonteCarlo
+from vgc.behaviour.BattlePolicies import RandomPlayer
 from vgc.behaviour.TeamBuildPolicies import TerminalTeamBuilder, RandomTeamBuilder
 from vgc.behaviour.TeamSelectionPolicies import FirstEditionTeamSelectionPolicy
 from vgc.competition.Competitor import Competitor
+from vgc.behaviour.myPolicy import MyPolicy
+from vgc.behaviour.otherPolicies import MyMinimax, MyMinimaxWithAlphaBetaKiller, MCTS_MS
 
 class MyCompetitor0(Competitor): 
 
@@ -108,83 +109,12 @@ class MyCompetitor3(Competitor):
     def battle_policy(self) -> BattlePolicy:
         return self._battle_policy
     
+
 class MyCompetitor4(Competitor): 
 
     def __init__(self, name: str = "My Example"):
         self._name = name
-        self._battle_policy = MyMinimaxWithAlphaBetaSortedKiller()
-        self._team_selection_policy = FirstEditionTeamSelectionPolicy()
-        self._team_build_policy = RandomTeamBuilder()
-
-    @property
-    def name(self):
-        return self._name
-
-    @property
-    def team_build_policy(self) -> TeamBuildPolicy:
-        return self._team_build_policy
-
-    @property
-    def team_selection_policy(self) -> TeamSelectionPolicy:
-        return self._team_selection_policy
-
-    @property
-    def battle_policy(self) -> BattlePolicy:
-        return self._battle_policy
-    
-class MyCompetitor5(Competitor): 
-
-    def __init__(self, name: str = "My Example"):
-        self._name = name
-        self._battle_policy = MyMinimaxWithAlphaBetaKillertransposition()
-        self._team_selection_policy = FirstEditionTeamSelectionPolicy()
-        self._team_build_policy = RandomTeamBuilder()
-
-    @property
-    def name(self):
-        return self._name
-
-    @property
-    def team_build_policy(self) -> TeamBuildPolicy:
-        return self._team_build_policy
-
-    @property
-    def team_selection_policy(self) -> TeamSelectionPolicy:
-        return self._team_selection_policy
-
-    @property
-    def battle_policy(self) -> BattlePolicy:
-        return self._battle_policy
-    
-class MyCompetitor6(Competitor): 
-
-    def __init__(self, name: str = "My Example"):
-        self._name = name
         self._battle_policy = MCTS_MS()
-        self._team_selection_policy = FirstEditionTeamSelectionPolicy()
-        self._team_build_policy = RandomTeamBuilder()
-
-    @property
-    def name(self):
-        return self._name
-
-    @property
-    def team_build_policy(self) -> TeamBuildPolicy:
-        return self._team_build_policy
-
-    @property
-    def team_selection_policy(self) -> TeamSelectionPolicy:
-        return self._team_selection_policy
-
-    @property
-    def battle_policy(self) -> BattlePolicy:
-        return self._battle_policy
-    
-class MyCompetitor7(Competitor): 
-
-    def __init__(self, name: str = "My Example"):
-        self._name = name
-        self._battle_policy = AggressiveMinimaxWithMonteCarlo()
         self._team_selection_policy = FirstEditionTeamSelectionPolicy()
         self._team_build_policy = RandomTeamBuilder()
 
@@ -229,7 +159,7 @@ class ExampleCompetitor(Competitor):
     def battle_policy(self) -> BattlePolicy:
         return self._battle_policy
 
-
+"""
 class TerminalExampleCompetitor(ExampleCompetitor):
 
     def __init__(self, name: str = "TerminalPlayer"):
@@ -254,7 +184,7 @@ class TerminalExampleCompetitor(ExampleCompetitor):
     def battle_policy(self) -> BattlePolicy:
         return self._battle_policy
 
-
+"""
 """
 class GUIExampleCompetitor(ExampleCompetitor):
